requirements.txt
# Core LangChain packages
langchain==0.2.14
langchain-core==0.2.36
langchain-community==0.2.11

# Google Gemini integration
langchain-google-genai==1.0.10
langsmith==0.1.132

# Vector database
chromadb==0.5.5
tiktoken

# Environment variables
python-dotenv>=1.0.0

# Web framework
django>=4.2

# Optional utilities
GitPython
